pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py5"
dynamic = ["version"]
description = "Processing for CPython"
readme = "README.md"
license = ""
requires-python = ">3.8"
authors = [
    { name = "Jim Schmitz", email = "jim@ixora.io" },
]
keywords = [
    "IPython",
    "Jupyter",
    "Processing",
    "Widgets",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v2 (LGPLv2)",
    "Programming Language :: Java",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Artistic Software",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "autopep8>=2.0",
    "jpype1>=1.4",
    "line_profiler>=4.0",
    "numpy>=1.23",
    "pillow>=9.2",
    "pyobjc>=9.0;sys_platform==\"darwin\"",
    "requests>=2.28",
    "stackprinter>=0.2.4",
]

[project.optional-dependencies]
jupyter = [
    "py5jupyter>=0.2.0a0",
]

[project.scripts]
py5cmd = "py5_tools.tools.py5cmd:main"
py5translate-imported2module = "py5_tools.tools.py5translate_imported2module:main"
py5translate-module2imported = "py5_tools.tools.py5translate_module2imported:main"
py5translate-processingpy2imported = "py5_tools.tools.py5translate_processingpy2imported:main"
py5utils = "py5_tools.tools.py5utils:main"
run_sketch = "py5_tools.tools.run_sketch:main"

[project.urls]
"Bug Tracker" = "https://github.com/py5coding/py5generator/issues"
Documentation = "https://py5coding.org/"
Download = "https://pypi.org/project/py5"
Homepage = "https://py5coding.org/"
"Source Code" = "https://github.com/py5coding/py5"

[tool.hatch.version]
path = "py5/__init__.py"

[tool.hatch.build]
packages = [
    "/py5",
    "/py5_tools",
]
